Refactor custom random seed option
