--- src/core/config/custom_random/option.cpp
@@ -1,10 +1,0 @@
-#include "config/custom_random/option.h"
-
-#include <utility>
-
-#include "config/names_and_descriptions.h"
-
-namespace config {
-extern CommonOption<CustomRandomFlagType> const kCustomRandomFlagOpt{
-        names::kCustomRandom, descriptions::kDCustomRandom, std::make_pair(false, 47)};
-}  // namespace config
