--- src/core/config/custom_random/option.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include "config/common_option.h"
-#include "config/custom_random/type.h"
-
-namespace config {
-extern CommonOption<CustomRandomFlagType> const kCustomRandomFlagOpt;
-}  // namespace config
