+++ src/core/config/names.h
@@ -11,7 +11,7 @@ constexpr auto kSeparator = "separator";
 constexpr auto kHasHeader = "has_header";
 constexpr auto kEqualNulls = "is_null_equal_null";
 constexpr auto kThreads = "threads";
-constexpr auto kCustomRandom = "custom_random";
+constexpr auto kCustomRandom = "custom_random_seed";
 constexpr auto kError = "error";
 constexpr auto kPfdErrorMeasure = "pfd_error_measure";
 constexpr auto kAfdErrorMeasure = "afd_error_measure";
