+++ src/python_bindings/py_util/get_py_type.cpp
@@ -13,7 +13,7 @@
 #include "algorithms/md/hymd/hymd.h"
 #include "algorithms/metric/enums.h"
 #include "association_rules/ar_algorithm_enums.h"
-#include "config/custom_random/type.h"
+#include "config/custom_random_seed/type.h"
 #include "config/error_measure/type.h"
 #include "config/tabular_data/input_table_type.h"
 #include "config/tabular_data/input_tables_type.h"
@@ -78,7 +78,7 @@ py::tuple GetPyType(std::type_index type_index) {
             PyTypePair<size_t, kPyInt>,
             PyTypePair<long double, kPyFloat>,
             PyTypePair<std::size_t, kPyInt>,
-            PyTypePair<config::CustomRandomFlagType, kPyTuple>,
+            PyTypePair<config::CustomRandomSeedType, kPyInt>,
             PyTypePair<algos::metric::Metric, kPyStr>,
             PyTypePair<algos::metric::MetricAlgo, kPyStr>,
             PyTypePair<config::PfdErrorMeasureType, kPyStr>,
